Split SelectableDataSource into DataSource and SelectableDataSource
