+++ Pod/Core/Classes/Protocol/SelectableDataSource.swift
@@ -26,26 +26,7 @@ import Photos
 /**
 Selectable data source protocol. Based on index paths and coupled to Photos for now...
 */
-public protocol SelectableDataSource : NSObjectProtocol {
-    /**
-    Data source delegate (notify if any updates have been done, etc).
-    */
-    var delegate: SelectableDataDelegate? { get set }
-    
-    /**
-    Number of sections in the data source
-    */
-    var sections: Int { get }
-    
-    /**
-    Number of objects for a given section.
-    */
-    func numberOfObjectsInSection(section: Int) -> Int
-    /**
-    Get object at given index path.
-    */
-    func objectAtIndexPath(indexPath: NSIndexPath) -> PHObject
-    
+public protocol SelectableDataSource : DataSource {
     /**
     Does this data source allow multiple selection?. 
     true: Selection is allowed up to maxNumberOfSelections.
